fix: updated angular.json with a working configuration for conditional styles, updated angular to v14
the styles were moved to the assets folder for better consistency
