+++ frontend/package.json
@@ -1,92 +1,92 @@
 {
-  "name": "frontend",
-  "version": "0.0.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve --ssl --host=0.0.0.0",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular-material-components/datetime-picker": "^7.0.0",
-    "@angular-material-components/moment-adapter": "^7.0.0",
-    "@angular/animations": "^14.2.12",
-    "@angular/cdk": "^13.3.9",
-    "@angular/cli": "^14.2.10",
-    "@angular/common": "^14.2.12",
-    "@angular/compiler": "^14.2.12",
-    "@angular/core": "^14.2.12",
-    "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "^14.2.12",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "^14.2.12",
-    "@angular/platform-browser-dynamic": "^14.2.12",
-    "@angular/router": "^14.2.12",
-    "@ckeditor/ckeditor5-angular": "^1.2.3",
-    "@ckeditor/ckeditor5-build-classic": "^22.0.0",
-    "@ericblade/quagga2": "^1.3.1",
-    "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "@types/leaflet": "^1.7.4",
-    "@zxing/browser": "0.0.10",
-    "@zxing/library": "^0.18.6",
-    "@zxing/ngx-scanner": "^2.0.0",
-    "angular-calendar": "^0.28.16",
-    "angularx-flatpickr": "^6.4.0",
-    "angularx-qrcode": "^14.0.0",
-    "d3": "^7.4.2",
-    "date-fns": "^2.15.0",
-    "flatpickr": "^4.6.6",
-    "gent_styleguide": "^5.0.10",
-    "html-entities": "^2.3.2",
-    "jquery": "^3.5.1",
-    "leaflet": "^1.7.1",
-    "moment": "^2.29.1",
-    "moment-duration-format": "^2.3.2",
-    "ngx-markdown": "^14.0.0",
-    "ngx-print": "^1.2.0-beta.5",
-    "rxjs": "~6.5.5",
-    "tslib": "^2.2.0",
-    "zone.js": "~0.11.4"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.10",
-    "@angular-eslint/builder": "1.2.0",
-    "@angular-eslint/eslint-plugin": "1.2.0",
-    "@angular-eslint/eslint-plugin-template": "1.2.0",
-    "@angular-eslint/schematics": "^14.4.0",
-    "@angular-eslint/template-parser": "1.2.0",
-    "@angular/compiler-cli": "^14.2.12",
-    "@types/angular": "^1.8.0",
-    "@types/d3": "^7.1.0",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/moment-duration-format": "^2.2.2",
-    "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "4.3.0",
-    "@typescript-eslint/parser": "4.3.0",
-    "codelyzer": "^0.0.28",
-    "eslint": "^7.6.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsdoc": "30.7.6",
-    "eslint-plugin-prefer-arrow": "1.2.2",
-    "eslint-plugin-prettier": "^3.3.1",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.4",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "prettier": "^2.2.1",
-    "protractor": "~7.0.0",
-    "schematics-scss-migrate": "^1.3.13",
-    "ts-node": "~8.3.0",
-    "tslint": "^6.1.3",
-    "typescript": "~4.6.4"
-  }
+    "name": "frontend",
+    "version": "0.0.0",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve --ssl --host=0.0.0.0",
+        "build": "ng build",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular-material-components/datetime-picker": "^7.0.1",
+        "@angular-material-components/moment-adapter": "^7.0.0",
+        "@angular/animations": "^14.2.12",
+        "@angular/cdk": "^13.3.9",
+        "@angular/cli": "^14.2.10",
+        "@angular/common": "^14.2.12",
+        "@angular/compiler": "^14.2.12",
+        "@angular/core": "^14.2.12",
+        "@angular/flex-layout": "^13.0.0-beta.38",
+        "@angular/forms": "^14.2.12",
+        "@angular/material": "^13.3.9",
+        "@angular/platform-browser": "^14.2.12",
+        "@angular/platform-browser-dynamic": "^14.2.12",
+        "@angular/router": "^14.2.12",
+        "@ckeditor/ckeditor5-angular": "^1.2.3",
+        "@ckeditor/ckeditor5-build-classic": "^22.0.0",
+        "@ericblade/quagga2": "^1.8.2",
+        "@ngx-translate/core": "^13.0.0",
+        "@ngx-translate/http-loader": "^6.0.0",
+        "@types/leaflet": "^1.9.1",
+        "@zxing/browser": "0.0.10",
+        "@zxing/library": "^0.18.6",
+        "@zxing/ngx-scanner": "^2.0.1",
+        "angular-calendar": "^0.28.28",
+        "angularx-flatpickr": "^6.6.0",
+        "angularx-qrcode": "^14.0.0",
+        "d3": "^7.8.2",
+        "date-fns": "^2.29.3",
+        "flatpickr": "^4.6.13",
+        "gent_styleguide": "^5.1.4",
+        "html-entities": "^2.3.3",
+        "jquery": "^3.6.3",
+        "leaflet": "^1.9.3",
+        "moment": "^2.29.4",
+        "moment-duration-format": "^2.3.2",
+        "ngx-markdown": "^14.0.1",
+        "ngx-print": "^1.3.1",
+        "rxjs": "~6.5.5",
+        "tslib": "^2.5.0",
+        "zone.js": "^0.11.8"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "^14.2.10",
+        "@angular-eslint/builder": "1.2.0",
+        "@angular-eslint/eslint-plugin": "1.2.0",
+        "@angular-eslint/eslint-plugin-template": "1.2.0",
+        "@angular-eslint/schematics": "^14.4.0",
+        "@angular-eslint/template-parser": "1.2.0",
+        "@angular/compiler-cli": "^14.2.12",
+        "@types/angular": "^1.8.4",
+        "@types/d3": "^7.4.0",
+        "@types/jasmine": "~3.6.0",
+        "@types/jasminewd2": "~2.0.3",
+        "@types/moment-duration-format": "^2.2.3",
+        "@types/node": "^12.20.55",
+        "@typescript-eslint/eslint-plugin": "4.3.0",
+        "@typescript-eslint/parser": "4.3.0",
+        "codelyzer": "^0.0.28",
+        "eslint": "^7.32.0",
+        "eslint-config-prettier": "^8.6.0",
+        "eslint-plugin-import": "2.22.1",
+        "eslint-plugin-jsdoc": "30.7.6",
+        "eslint-plugin-prefer-arrow": "1.2.2",
+        "eslint-plugin-prettier": "^3.4.1",
+        "jasmine-core": "~3.6.0",
+        "jasmine-spec-reporter": "~5.0.0",
+        "karma": "~6.3.4",
+        "karma-chrome-launcher": "~3.1.0",
+        "karma-coverage-istanbul-reporter": "~3.0.2",
+        "karma-jasmine": "~4.0.0",
+        "karma-jasmine-html-reporter": "^1.7.0",
+        "prettier": "^2.8.4",
+        "protractor": "~7.0.0",
+        "schematics-scss-migrate": "^1.3.15",
+        "ts-node": "~8.3.0",
+        "tslint": "^6.1.3",
+        "typescript": "~4.6.4"
+    }
 }
