+++ frontend/src/app/miscellaneous/qrcode/qrcode.component.html
@@ -3,4 +3,4 @@
         <!-- NOTE: If you change the width here, also change the width in the scss file -->
         <qrcode *ngIf="userObs | async as user" [qrdata]="user.userId" [width]="256" [errorCorrectionLevel]="'M'"></qrcode>
     </div>
-  </div>
+</div>
