+++ frontend/src/assets/styles/local/styles.scss
@@ -1,0 +1,30 @@
+/* The development styles are an extension of the production styles. */
+@import '../production/styles';
+
+/* Make it clear that we're in the development environment. */
+.bg-primary {
+    background-color: #F1A42B !important;
+}
+
+.pageheader .navbar-default .nav-primary > li.active > a span {
+    color: #F1A42B !important;
+    border-color: #F1A42B !important;
+}
+
+.pageheader .navbar-default .nav-primary > li > a, .pageheader .navbar-default .nav-primary > li > a:hover, .pageheader .navbar-default .nav-primary > li > a:focus {
+    color: #F1A42B !important;
+    border-color: #F1A42B !important;
+    text-decoration-color: #F1A42B !important;
+}
+
+header.header hr::after {
+    border-color: #F1A42B !important;
+}
+
+header.header hr::before {
+    border-color: #F1A42B !important;
+}
+
+.cs--cyan .theme-footer {
+    background-color: #F1A42B !important;
+}
