+++ frontend/src/assets/styles/production/styles.scss
@@ -1,7 +1,9 @@
 // Setting up gent_styleguide (cfr. https://stijlgids.stad.gent/v4/index.html)
-@use '@angular/material' as mat;
 $styleguide-dir: "" !default;
-@import "./node_modules/gent_styleguide/build/styleguide/sass/main_cli";
+
+@use '../../../../node_modules/@angular/material/index' as mat;
+
+@import "../../../../node_modules/gent_styleguide/build/styleguide/sass/main_cli";
 
 // Custom Theming for Angular Material
 // For more information: https://material.angular.io/guide/theming
@@ -11,27 +13,21 @@ $styleguide-dir: "" !default;
 // Include the common styles for Angular Material. We include this here so that you only
 // have to load a single css file for Angular Material in your app.
 // Be sure that you only ever include this mixin once!
-
 @include mat.core();
 
 // import our custom theme
-@import "./theme.scss";
-
-// import custom component themes
-@import "./custom-component-themes.scss";
+@import 'src/assets/styles/theme';
 
 /* angular-cli file: src/styles.css */
-@import "../node_modules/angular-calendar/css/angular-calendar.css";
+@import '../../../../node_modules/angular-calendar/css/angular-calendar.css';
 
 // import leaflet styling
-@import "../node_modules/leaflet/dist/leaflet.css";
+@import '../../../../node_modules/leaflet/dist/leaflet.css';
 
 // Include theme styles for core and each component used in your app.
 // Alternatively, you can import and @include the theme mixins for each component
 // that you are using.
-
 @include mat.all-component-themes($theming-material-components-theme);
-@include custom-components-theme($theming-material-components-theme);
 
 // Include the dark color styles inside of a block with a CSS class. You can make this
 // CSS class whatever you want. In this example, any component inside of an element with
