+++ frontend/src/assets/styles/staging/styles.scss
@@ -1,0 +1,2 @@
+/* The staging styles are an extension of the production styles. */
+@import '../production/styles';
