+++ frontend/src/assets/styles/theme.scss
@@ -1,10 +1,10 @@
-// src/theme.scss
-@use '@angular/material' as mat;
+// src/theme.styles
+@use '../../../node_modules/@angular/material/index' as mat;
 
-
-// 👇 1. Default theme
+// 2. Default theme
 $theming-material-components-primary: mat.define-palette(mat.$blue-palette, 800);
 $theming-material-components-accent: mat.define-palette(mat.$pink-palette, A200, A100, A400);
+
 // The warn palette is optional (defaults to red).
 $theming-material-components-warn: mat.define-palette(mat.$red-palette);
 
@@ -17,22 +17,7 @@ $theming-material-components-theme: mat.define-light-theme((
     warn: $theming-material-components-warn)
 ));
 
-
-// 👇 2. Define an alternate dark theme.
-//$dark-primary: mat-palette($mat-blue-grey);
-//$dark-accent: mat-palette($mat-amber, A200, A100, A400);
-//$dark-warn: mat-palette($mat-deep-orange);
-//$dark-theme: mat-dark-theme(
-//    (
-//      color: (
-//        primary: $dark-primary,
-//        accent: $dark-accent,
-//        warn: $dark-warn,
-//      ),
-//    )
-//);
-
-// 👇 3. Define a custom heading typography config
+// 2. Define a custom heading typography config
 $heading-font-family: "'PannoTextMedium',Arial, sans-serif";
 
 $typography: mat.define-typography-config(
