--- frontend/src/custom-component-themes.scss
@@ -1,5 +1,0 @@
-// import custom components themes
-
-// you only have to add additional components here (instead of in every theme class)
-@mixin custom-components-theme($theme) {
-}
