--- frontend/src/development.scss
@@ -1,25 +1,0 @@
-.bg-primary {
-  background-color: #F1A42B !important;
-}
-.pageheader .navbar-default .nav-primary>li.active>a span {
-  color: #F1A42B !important;
-  border-color: #F1A42B !important;
-}
-.pageheader .navbar-default .nav-primary>li>a, .pageheader .navbar-default .nav-primary>li>a:hover, .pageheader .navbar-default .nav-primary>li>a:focus {
-  color: #F1A42B !important;
-  border-color: #F1A42B !important;
-  text-decoration-color: #F1A42B !important;
-}
-
-header.header hr::after {
-  border-color: #F1A42B !important;
-}
-
-
-header.header hr::before {
-  border-color: #F1A42B !important;
-}
-
-.cs--cyan .theme-footer {
-  background-color: #F1A42B !important;
-}
