--- frontend/src/extra-styling-placeholder.scss
@@ -1,1 +1,0 @@
-// This file should be empty. In production environment this file is used to replace the development.scss styling.
