+++ frontend/src/index.html
@@ -1,4 +1,4 @@
-<!doctype html>
+
 <html lang="en">
 <head>
   <meta charset="utf-8">
@@ -9,7 +9,6 @@
   <link rel="icon" type="image/x-icon" href="/assets/images/favicon-32x32.png">
   <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500&display=swap" rel="stylesheet">
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-
 </head>
 <body id="body">
   <app-root></app-root>
