+++ frontend/src/polyfills.ts
@@ -23,12 +23,12 @@
  */
 
 /**
- * By default, zone.js will patch all possible macroTask and DomEvents
+ * By default, zone.scripts will patch all possible macroTask and DomEvents
  * user can disable parts of macroTask/DomEvents patch by setting following flags
- * because those flags need to be set before `zone.js` being loaded, and webpack
+ * because those flags need to be set before `zone.scripts` being loaded, and webpack
  * will put import in the top of bundle, so user need to create a separate file
  * in this directory (for example: zone-flags.ts), and put the following flags
- * into that file, and then add the following code before importing zone.js.
+ * into that file, and then add the following code before importing zone.scripts.
  * import './zone-flags';
  *
  * The flags allowed in zone-flags.ts are listed here.
@@ -39,8 +39,8 @@
  * (window as any).__Zone_disable_on_property = true; // disable patch onProperty such as onclick
  * (window as any).__zone_symbol__UNPATCHED_EVENTS = ['scroll', 'mousemove']; // disable patch specified eventNames
  *
- *  in IE/Edge developer tools, the addEventListener will also be wrapped by zone.js
- *  with the following flag, it will bypass `zone.js` patch for IE/Edge
+ *  in IE/Edge developer tools, the addEventListener will also be wrapped by zone.scripts
+ *  with the following flag, it will bypass `zone.scripts` patch for IE/Edge
  *
  *  (window as any).__Zone_enable_cross_context_check = true;
  *
