+++ frontend/src/test.ts
@@ -1,4 +1,4 @@
-// This file is required by karma.conf.js and loads recursively all the .spec and framework files
+// This file is required by karma.conf.scripts and loads recursively all the .spec and framework files
 
 import 'zone.js/testing';
 import { getTestBed } from '@angular/core/testing';
